requirements.txt
streamlit==1.19.0
plotly==5.9.0
pandas==1.4.4
numpy==1.21.5
folium==0.14.0
matplotlib==3.5.2
matplotlib-inline==0.1.6
haversine==2.8.0
streamlit-folium==0.11.1
Pillow==9.2.0